// src/pages/mbb-documents/internal-inbound/interfaces/internal-inbound-mbb.interface.ts
export interface InternalInboundMbbDocument {
  _id: string;
  reg_num: string;
  reg_date: string;
  journal_index: string;
  original_num: string;
  original_date: string;
  doc_num: string;
  doc_date: string;
  doc_type: string;
  organization: string;
  content: string;
  assignee: string;
  resolution: string;
  deadline: string;
  status: string;
  reply_order_date: string;
}